// libs/shared/modules/src/data/models/domain/framework.ts
/**
 * Anecdotes API - Integration Environment
 * Anecdotes API's
 *
 * OpenAPI spec version: 0.0.3
 * Contact: apiteam@anecdotes.ai
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Audit } from './audit';

export interface ExcludePlugin {
  [key: string]: string[];
}

export interface Framework {
  /**
   * whether this framework applicable or not
   */
  is_applicable?: boolean;
  /**
   * the framework id
   */
  framework_id?: string;
  /**
   * the framework name
   */
  framework_name?: string;
  /**
   * The current status for this framework (AVAILABLE/COMINGSOON)
   */
  framework_status?: string;
  /**

   * The description of this framework
   */
  framework_description?: string;
  /**

   * number of controls that related to this framework
   */
  number_of_related_controls?: number;
  framework_reference_field_name?: string;
  /**
   * list of framework references
   */
  framework_references?: Array<string>;
  /**
   * whether this framework auditable or not
   */
  framework_auditable?: boolean;
  framework_current_audit?: Audit;

  /**
  * The overriden framework id
  */
  framework_icon_id?: string;
  /**
   * Excluded plugins from framework
   */
  framework_excluded_plugins?: ExcludePlugin;
  
  freeze?: boolean;
  related_controls_snapshots?: string[];
  
  snapshot_id?: string;

  is_snapshot?: boolean;
}