// libs/shared/modules/src/data/models/domain/audit.ts
/**
 * Anecdotes API
 * Anecdotes API's
 *
 * OpenAPI spec version: 0.0.3
 * Contact: apiteam@anecdotes.ai
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface Audit { 
    id?: string;
    framework_id?: string;
    product_name?: string;
    update_by?: string;
    audit_date?: Date;
    update_time?: Date;
    framework_fields?: any;
}